+++ includes/ui/mybooking-plugin-ui-pages.php
@@ -6,44 +6,6 @@
   {
 
      /**
-      * Builds and loads the products page
-      *
-      * @param page the page index
-      * @param limit the number of products in the page
-      */
-     public function products($page, $limit) {
-
-				  $registry = Mybooking_Registry::getInstance();
-          $url = $registry->mybooking_rent_plugin_navigation_products_url ? $registry->mybooking_rent_plugin_navigation_products_url : 'products';
-          $offset = ($page - 1) * $limit;
-
-          // Call the API 
-          $api_client = new MybookingApiClient($registry->mybooking_rent_plugin_api_url_prefix,
-          	                                   $registry->mybooking_rent_plugin_api_key);
-          $data =$api_client->get_products($offset, $limit);
-          if ( $data == null) {
-          	$this->routes_not_found();
-          }
-
-          // Calculate pagination
-          $total_pages = ceil($data->total / $data->limit);
-          $current_page = floor($data->offset / $data->limit) + 1; 
-          $pagination = new MyBookingUIPagination();          
-          $pages = $pagination->pages($total_pages, $current_page);
-
-          // Build the pages
-          $data = array('data' => $data,
-          	            'total_pages' => $total_pages,
-          	            'current_page' => $current_page,
-          	            'pages' => $pages,
-                        'url' => $url);
-
-          mybooking_engine_get_template('mybooking-plugin-routes-products.php', $data);
-          die;    
-
-     }
-
-     /**
       * Builds and loads the product page
       *
       * @param code The product code